+++ packages/editor/src/components/collab-sidebar/add-comment.js
@@ -7,28 +7,27 @@ import {
 	__experimentalHStack as HStack,
 	__experimentalVStack as VStack,
 } from '@wordpress/components';
-import { store as blockEditorStore } from '@wordpress/block-editor';
+import {
+	store as blockEditorStore,
+	privateApis as blockEditorPrivateApis,
+} from '@wordpress/block-editor';
 import { isUnmodifiedDefaultBlock } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
+import { unlock } from '../../lock-unlock';
 import CommentAuthorInfo from './comment-author-info';
 import CommentForm from './comment-form';
+import { focusCommentThread } from './utils';
 
-/**
- * Renders the UI for adding a comment in the Gutenberg editor's collaboration sidebar.
- *
- * @param {Object}   props                     - The component props.
- * @param {Function} props.onSubmit            - A callback function to be called when the user submits a comment.
- * @param {boolean}  props.showCommentBoard    - The function to edit the comment.
- * @param {Function} props.setShowCommentBoard - The function to delete the comment.
- * @return {React.ReactNode} The rendered comment input UI.
- */
+const { useBlockElement } = unlock( blockEditorPrivateApis );
+
 export function AddComment( {
 	onSubmit,
 	showCommentBoard,
 	setShowCommentBoard,
+	commentSidebarRef,
 } ) {
 	const { clientId, blockCommentId, isEmptyDefaultBlock } = useSelect(
 		( select ) => {
@@ -41,8 +40,10 @@ export function AddComment( {
 					? isUnmodifiedDefaultBlock( selectedBlock )
 					: false,
 			};
-		}
+		},
+		[]
 	);
+	const blockElement = useBlockElement( clientId );
 
 	if (
 		! showCommentBoard ||
@@ -52,8 +53,6 @@ export function AddComment( {
 	) {
 		return null;
 	}
-
-	const commentLabel = __( 'New Comment' );
 
 	return (
 		<VStack
@@ -66,14 +65,16 @@ export function AddComment( {
 				<CommentAuthorInfo />
 			</HStack>
 			<CommentForm
-				onSubmit={ ( inputComment ) => {
-					onSubmit( { content: inputComment } );
+				onSubmit={ async ( inputComment ) => {
+					const { id } = await onSubmit( { content: inputComment } );
+					focusCommentThread( id, commentSidebarRef.current );
 				} }
 				onCancel={ () => {
 					setShowCommentBoard( false );
+					blockElement?.focus();
 				} }
 				submitButtonText={ _x( 'Comment', 'Add comment button' ) }
-				labelText={ commentLabel }
+				labelText={ __( 'New Comment' ) }
 			/>
 		</VStack>
 	);