+++ packages/editor/src/components/collab-sidebar/index.js
@@ -2,16 +2,13 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { useSelect, useDispatch, subscribe } from '@wordpress/data';
+import { useSelect, useDispatch } from '@wordpress/data';
 import { __experimentalVStack as VStack } from '@wordpress/components';
 import { useState, useRef } from '@wordpress/element';
 import { useViewportMatch } from '@wordpress/compose';
 import { comment as commentIcon } from '@wordpress/icons';
-import { store as noticesStore } from '@wordpress/notices';
-import { store as coreStore } from '@wordpress/core-data';
 import { store as blockEditorStore } from '@wordpress/block-editor';
 import { store as interfaceStore } from '@wordpress/interface';
-import { decodeEntities } from '@wordpress/html-entities';
 
 /**
  * Internal dependencies
@@ -21,10 +18,15 @@ import { collabHistorySidebarName, collabSidebarName } from './constants';
 import { Comments } from './comments';
 import { AddComment } from './add-comment';
 import { store as editorStore } from '../../store';
-import AddCommentButton from './comment-button';
+import AddCommentMenuItem from './comment-menu-item';
 import CommentAvatarIndicator from './comment-indicator-toolbar';
 import { useGlobalStylesContext } from '../global-styles-provider';
-import { useBlockComments } from './hooks';
+import {
+	useBlockComments,
+	useBlockCommentsActions,
+	useEnableFloatingSidebar,
+} from './hooks';
+import { focusCommentThread } from './utils';
 
 function CollabSidebarContent( {
 	showCommentBoard,
@@ -33,139 +35,7 @@ function CollabSidebarContent( {
 	comments,
 	commentSidebarRef,
 } ) {
-	const { createNotice } = useDispatch( noticesStore );
-	const { saveEntityRecord, deleteEntityRecord } = useDispatch( coreStore );
-	const { updateBlockAttributes } = useDispatch( blockEditorStore );
-	const { currentPostId, getSelectedBlockClientId, getBlockAttributes } =
-		useSelect( ( select ) => {
-			const { getCurrentPostId } = select( editorStore );
-			return {
-				getSelectedBlockClientId:
-					select( blockEditorStore ).getSelectedBlockClientId,
-				getBlockAttributes:
-					select( blockEditorStore ).getBlockAttributes,
-				currentPostId: getCurrentPostId(),
-			};
-		}, [] );
-
-	const onError = ( error ) => {
-		const errorMessage =
-			error.message && error.code !== 'unknown_error'
-				? decodeEntities( error.message )
-				: __( 'An error occurred while performing an update.' );
-		createNotice( 'error', errorMessage, {
-			type: 'snackbar',
-			isDismissible: true,
-		} );
-	};
-
-	const addNewComment = async ( { content, parent } ) => {
-		try {
-			const savedRecord = await saveEntityRecord(
-				'root',
-				'comment',
-				{
-					post: currentPostId,
-					content,
-					comment_type: 'block_comment',
-					comment_approved: 0,
-					parent: parent || 0,
-				},
-				{ throwOnError: true }
-			);
-
-			// If it's a main comment, update the block attributes with the comment id.
-			if ( ! parent && savedRecord?.id ) {
-				const metadata = getBlockAttributes(
-					getSelectedBlockClientId()
-				)?.metadata;
-				updateBlockAttributes( getSelectedBlockClientId(), {
-					metadata: {
-						...metadata,
-						commentId: savedRecord.id,
-					},
-				} );
-			}
-
-			createNotice(
-				'snackbar',
-				parent
-					? __( 'Reply added successfully.' )
-					: __( 'Comment added successfully.' ),
-				{
-					type: 'snackbar',
-					isDismissible: true,
-				}
-			);
-		} catch ( error ) {
-			onError( error );
-		}
-	};
-
-	const onEditComment = async ( { id, content, status } ) => {
-		const messageType = status ? status : 'updated';
-		const messages = {
-			approved: __( 'Comment marked as resolved.' ),
-			hold: __( 'Comment reopened.' ),
-			updated: __( 'Comment updated.' ),
-		};
-
-		try {
-			await saveEntityRecord(
-				'root',
-				'comment',
-				{
-					id,
-					content,
-					status,
-				},
-				{ throwOnError: true }
-			);
-			createNotice(
-				'snackbar',
-				messages[ messageType ] ?? __( 'Comment updated.' ),
-				{
-					type: 'snackbar',
-					isDismissible: true,
-				}
-			);
-		} catch ( error ) {
-			onError( error );
-		}
-	};
-
-	const onCommentDelete = async ( comment ) => {
-		try {
-			await deleteEntityRecord(
-				'root',
-				'comment',
-				comment.id,
-				undefined,
-				{
-					throwOnError: true,
-				}
-			);
-
-			if ( ! comment.parent ) {
-				const metadata = getBlockAttributes(
-					getSelectedBlockClientId()
-				)?.metadata;
-				updateBlockAttributes( getSelectedBlockClientId(), {
-					metadata: {
-						...metadata,
-						commentId: undefined,
-					},
-				} );
-			}
-
-			createNotice( 'snackbar', __( 'Comment deleted successfully.' ), {
-				type: 'snackbar',
-				isDismissible: true,
-			} );
-		} catch ( error ) {
-			onError( error );
-		}
-	};
+	const { onCreate, onEdit, onDelete } = useBlockCommentsActions();
 
 	return (
 		<div
@@ -175,15 +45,16 @@ function CollabSidebarContent( {
 		>
 			<VStack role="list" spacing="3">
 				<AddComment
-					onSubmit={ addNewComment }
+					onSubmit={ onCreate }
 					showCommentBoard={ showCommentBoard }
 					setShowCommentBoard={ setShowCommentBoard }
+					commentSidebarRef={ commentSidebarRef }
 				/>
 				<Comments
 					threads={ comments }
-					onEditComment={ onEditComment }
-					onAddReply={ addNewComment }
-					onCommentDelete={ onCommentDelete }
+					onEditComment={ onEdit }
+					onAddReply={ onCreate }
+					onCommentDelete={ onDelete }
 					showCommentBoard={ showCommentBoard }
 					setShowCommentBoard={ setShowCommentBoard }
 					commentSidebarRef={ commentSidebarRef }
@@ -198,8 +69,8 @@ function CollabSidebarContent( {
  */
 export default function CollabSidebar() {
 	const [ showCommentBoard, setShowCommentBoard ] = useState( false );
-	const { enableComplementaryArea } = useDispatch( interfaceStore );
 	const { getActiveComplementaryArea } = useSelect( interfaceStore );
+	const { enableComplementaryArea } = useDispatch( interfaceStore );
 	const isLargeViewport = useViewportMatch( 'medium' );
 	const commentSidebarRef = useRef( null );
 
@@ -220,34 +91,15 @@ export default function CollabSidebar() {
 			: null;
 	}, [] );
 
-	const openCollabBoard = () => {
-		setShowCommentBoard( true );
-		enableComplementaryArea( 'core', collabHistorySidebarName );
-	};
-
 	const { resultComments, unresolvedSortedThreads, totalPages } =
 		useBlockComments( postId );
+	useEnableFloatingSidebar( resultComments.length > 0 );
 
 	const hasMoreComments = totalPages && totalPages > 1;
 
 	// Get the global styles to set the background color of the sidebar.
 	const { merged: GlobalStyles } = useGlobalStylesContext();
 	const backgroundColor = GlobalStyles?.styles?.color?.background;
-
-	if ( 0 < resultComments.length ) {
-		const unsubscribe = subscribe( () => {
-			const activeSidebar = getActiveComplementaryArea( 'core' );
-
-			if ( ! activeSidebar ) {
-				enableComplementaryArea( 'core', collabSidebarName );
-				unsubscribe();
-			}
-		} );
-	}
-
-	const AddCommentComponent = blockCommentId
-		? CommentAvatarIndicator
-		: AddCommentButton;
 
 	// Find the current thread for the selected block.
 	const currentThread = blockCommentId
@@ -259,13 +111,36 @@ export default function CollabSidebar() {
 		return null;
 	}
 
+	async function openTheSidebar() {
+		enableComplementaryArea( 'core', collabHistorySidebarName );
+		const activeArea = await getActiveComplementaryArea( 'core' );
+
+		// Move focus to the target element after the sidebar has opened.
+		if (
+			[ collabHistorySidebarName, collabSidebarName ].includes(
+				activeArea
+			)
+		) {
+			setShowCommentBoard( ! blockCommentId );
+			focusCommentThread(
+				blockCommentId,
+				commentSidebarRef.current,
+				// Focus a comment thread when there's a selected block with a comment.
+				! blockCommentId ? 'textarea' : undefined
+			);
+		}
+	}
+
 	return (
 		<>
-			<AddCommentComponent
-				onClick={ openCollabBoard }
-				thread={ currentThread }
-				hasMoreComments={ hasMoreComments }
-			/>
+			{ blockCommentId && (
+				<CommentAvatarIndicator
+					thread={ currentThread }
+					hasMoreComments={ hasMoreComments }
+					onClick={ openTheSidebar }
+				/>
+			) }
+			<AddCommentMenuItem onClick={ openTheSidebar } />
 			<PluginSidebar
 				identifier={ collabHistorySidebarName }
 				// translators: Comments sidebar title