+++ test/e2e/specs/widgets/editing-widgets.spec.js
@@ -162,6 +162,7 @@ test.describe( 'Widgets screen', () => {
 
 		const paragraphBlock = inlineQuickInserter.getByRole( 'option', {
 			name: 'Paragraph',
+			exact: true,
 		} );
 		await paragraphBlock.click();
 
@@ -214,7 +215,7 @@ test.describe( 'Widgets screen', () => {
 		await page.keyboard.type( 'Heading' );
 
 		await inlineQuickInserter
-			.getByRole( 'option', { name: 'Heading' } )
+			.getByRole( 'option', { name: 'Heading', exact: true } )
 			.click();
 
 		await expect(
@@ -708,7 +709,10 @@ class WidgetsScreen {
 			} )
 			.fill( blockName );
 
-		return blockLibrary.getByRole( 'option', { name: blockName } );
+		return blockLibrary.getByRole( 'option', {
+			name: blockName,
+			exact: true,
+		} );
 	};
 
 	saveWidgets = async () => {