+++ packages/icons/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Remove `edit` and `warning` icons, which were aliases to `pencil` and `cautionFilled`, respectively. Switch to SVG icons as canonical source format, letting the system auto-generate the React elements and index. ([#71878](https://github.com/WordPress/gutenberg/pull/71878)).
 
+### Enhancements
+
+- Update `check` to be more optically balanced. ([#72101](https://github.com/WordPress/gutenberg/pull/72101))
+
 ## 10.32.0 (2025-10-01)
 
 ## 10.31.0 (2025-09-17)